attr: introduce Union type to reduce the number of functions in the trait

+++ crates/ap-storage-partition/src/dir.rs
@@ -25,16 +25,13 @@ impl DirIterator for PartitionDir<'_> {
             return Ok(None);
         }
 
-        let partition: Partition = self
-            .file
-            .disk
-            .read_object(self.file.offset + 0x1be + self.pos * 0x10)?;
+        let partition: Partition = self.file.disk.read_object(self.file.offset + 0x1be + self.pos * 0x10)?;
         let mut writer = SliceWriter(name, 0);
         core::write!(&mut writer, "part-{}", self.pos)?;
 
         let typ = if partition.typ == 0 || partition.size == 0 || writer.1 == 0 {
             FileType::Unknown
-        } else if let Ok(child) = self.file.open(self.pos as u64) {
+        } else if let Ok(child) = self.file.open(self.pos) {
             child.ftype()
         } else {
             FileType::Unknown
